Valider skjema for selvbestemt IM
