+++ api/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/api/response/RedisTimeoutResponse.kt
@@ -11,6 +11,36 @@ import java.util.UUID
 
 @Serializable
 @OptIn(ExperimentalSerializationApi::class)
+data class UkjentErrorResponse(
+    val inntektsmeldingTypeId: UUID
+) {
+    @EncodeDefault
+    val error = "Ukjent feil."
+}
+
+@Serializable
+@OptIn(ExperimentalSerializationApi::class)
+data class JsonErrorResponse(
+    // TODO slett etter endring i frontend
+    val forespoerselId: String? = null,
+    val inntektsmeldingTypeId: UUID? = null
+) {
+    @EncodeDefault
+    val error = "Feil under serialisering."
+}
+
+@Serializable
+@OptIn(ExperimentalSerializationApi::class)
+data class ValideringErrorResponse(
+    val inntektsmeldingTypeId: UUID,
+    val valideringsfeil: Set<String>
+) {
+    @EncodeDefault
+    val error = "Feil under validering."
+}
+
+@Serializable
+@OptIn(ExperimentalSerializationApi::class)
 data class RedisTimeoutResponse(
     // TODO slett etter endring i frontend
     val uuid: UUID? = null,
